+++ src/lib.rs
@@ -32,10 +32,16 @@ pub use worker::*;
 use fragile::Fragile;
 use once_cell::sync::OnceCell;
 
-static APP: OnceCell<Fragile<gtk::Application>> = OnceCell::new();
+static APP: OnceCell<Fragile<Application>> = OnceCell::new();
 
 pub use gtk;
 pub use gtk::glib::Sender;
+
+#[cfg(feature = "libadwaita")]
+type Application = adw::Application;
+
+#[cfg(not(feature = "libadwaita"))]
+type Application = gtk::Application;
 
 #[cfg(feature = "tokio-rt")]
 #[cfg_attr(doc, doc(cfg(feature = "tokio-rt")))]
@@ -49,7 +55,7 @@ pub use async_trait::async_trait;
 ///
 /// This function panics if [`RelmApp::new`] wasn't called before
 /// or this function is not called on the thread that also called [`RelmApp::new`].
-pub fn gtk_application() -> gtk::Application {
+pub fn gtk_application() -> Application {
     APP.get()
         .expect("The gloabl gtk application hasn't been initialized yet")
         .try_get()